Remove old sidebar components
The old sidebar is deprecated, and many of its components are not
used anywhere else. These have all been removed.
